Depend on QuickCheck-2.12 or later
I'm not sure if text-show-instances is affected by
nick8325/quickcheck#201 or not. But text-show definitely was (see
RyanGlScott/text-show#36), and since text-show-instances defines some
of the same Arbitrary instances for data types in Foreign.C.Types
that have different underlying representations based on OS and
architecture, it's better to be safe than sorry.