Depend on quickcheck-instances-0.3.19+
That version of quickcheck-instances introduces an Arbitrary instance
for MaybeT. This doesn't technically pose a problem at the moment,
since text-show-instances' own Arbitrary instance for MaybeT happens
not to collide with that of quickcheck-instances. But it's possible
that this could change in the future, so let's be safe and just use
the one from quickcheck-instances.