Pare down TH tests
Instances.Language.Haskell.TH is an absurdly complex swamp of CPP.
But it doesn't really deserve to be this way, especially since 99%
of the instances in that module exist only to test TH-derived
TextShow instances, which just isn't worth the maintenance burden.
This simplifies the module down to only test the instances that are
implemented by hand.